+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - (Feature) (Analytics) Enable TLS and Service
 - (Feature) (ML) Unify Integration Sidecar
 - (Feature) (Analytics) Metadata
+- (Feature) (Analytics) StatefulSet
 
 ## [1.2.40](https://github.com/arangodb/kube-arangodb/tree/1.2.40) (2024-04-10)
 - (Feature) Add Core fields to the Scheduler Container Spec